+++ th2_data_services/utils/event_utils.py
@@ -1,3 +1,7 @@
+from typing import Callable, Dict, List, Tuple, Set, Union
+from datetime import datetime
+from th2_data_services.utils import misc_utils
+from th2_data_services import EVENT_FIELDS_RESOLVER, MESSAGE_FIELDS_RESOLVER
 import json
 from th2_data_services.utils import misc_utils
 from typing import Callable, Dict, List, Tuple, Set, Union, Optional
@@ -8,8 +12,10 @@ from tabulate import tabulate
 # TODO -
 #   1. events: List[Dict] -- should be Iterable[Th2Event], where Th2Event = Dict. It can be changed in the future
 #   2. IDEA - it's difficult to understand what you will get when read some functions. I think add examples
+#   3. Rename all docstrings word Gets to Returns. That's different things.
 
 
+# STREAMING
 def get_category_frequencies(
     events: List[Dict], categories: List[str], categorizer: Callable, aggregation_level: str = "seconds"
 ) -> List[List[str]]:
@@ -43,7 +49,7 @@ def get_category_frequencies(
         events,
         categories,
         categorizer,
-        lambda e: e["startTimestamp"]["epochSecond"],
+        lambda e: EVENT_FIELDS_RESOLVER.get_start_timestamp(e)["epochSecond"],
         aggregation_level=aggregation_level,
     )
 
@@ -78,7 +84,7 @@ def get_type_frequencies(events: List[Dict], types: List[str], aggregation_level
             ...
         ]
     """
-    return get_category_frequencies(events, types, lambda e: e["eventType"], aggregation_level)
+    return get_category_frequencies(events, types, lambda e: EVENT_FIELDS_RESOLVER.get_type(e), aggregation_level)
 
 
 # USEFUL
@@ -111,7 +117,7 @@ def get_category_totals(
     for event in events:
         category = categorizer(event)
         if not ignore_status:
-            status = " [ok]" if event["successful"] else " [fail]"
+            status = " [ok]" if EVENT_FIELDS_RESOLVER.get_status(event) else " [fail]"
             category += status
         event_categories[category] += 1
 
@@ -120,7 +126,9 @@ def get_category_totals(
 
 # USEFUL
 # STREAMING
-# TODO - NOT-READY -- event["attachedMessageIds"] should be updated by resolver
+# example
+# eu.get_attached_messages_totals(d1)
+# defaultdict(<class 'int'>, {'envtn2_msfix5:first': 25262, 'envtn2_jpmfix1:second': 1702, 'env2_gscofixg2:second': 1702,...
 def get_attached_messages_totals(events: List[Dict]) -> Dict[str, int]:
     """Returns dictionary quantities of messages attached to events for each stream.
 
@@ -136,7 +144,7 @@ def get_attached_messages_totals(events: List[Dict]) -> Dict[str, int]:
     """
     streams = defaultdict(int)
     for event in events:
-        for message_id in event["attachedMessageIds"]:
+        for message_id in EVENT_FIELDS_RESOLVER.get_attached_messages_ids(event):
             key = message_id[: message_id.rindex(":")]
             streams[key] += 1
 
@@ -145,7 +153,6 @@ def get_attached_messages_totals(events: List[Dict]) -> Dict[str, int]:
 
 # USEFUL
 # STREAMING
-# TODO - NOT-READY -- event["attachedMessageIds"] should be updated by resolver
 def get_attached_message_ids(events: List[Dict]) -> Set[str]:
     """Returns the set of unique message IDs linked to all events.
 
@@ -163,12 +170,11 @@ def get_attached_message_ids(events: List[Dict]) -> Set[str]:
               ...
             }
     """
-    return set(message_id for event in events for message_id in event["attachedMessageIds"])
+    return set(message_id for event in events for message_id in EVENT_FIELDS_RESOLVER.get_attached_messages_ids(event))
 
 
 # USEFUL
 # STREAMING
-# TODO - NOT-READY -- event["parentEventId, eventId"] should be updated by resolver
 # TODO - It returns only parent events that not present in the events.
 #   Perhaps we need to find better name
 # we use something similar in our EventTree
@@ -192,8 +198,8 @@ def get_prior_parent_ids(events: List[Dict]) -> Set[str]:
     all_event_ids = set()
     parent_ids = set()
     for event in events:
-        parent_id = event["parentEventId"]
-        event_id = event["eventId"]
+        parent_id = EVENT_FIELDS_RESOLVER.get_parent_id(event)
+        event_id = EVENT_FIELDS_RESOLVER.get_id(event)
         if parent_id is not None and not parent_id in all_event_ids:
             parent_ids.add(parent_id)
         if event_id in parent_ids:
@@ -205,7 +211,6 @@ def get_prior_parent_ids(events: List[Dict]) -> Set[str]:
 
 # USEFUL
 # NOT STREAMING - keeps events
-# TODO - NOT-READY -- event["attachedMessageIds"] should be updated by resolver
 # BE AWARE!! - it can take up all your memory
 # O(N*M)
 def get_attached_message_ids_index(events: List[Dict]) -> Dict[str, list]:
@@ -234,7 +239,7 @@ def get_attached_message_ids_index(events: List[Dict]) -> Dict[str, list]:
     """
     result = defaultdict(list)
     for event in events:
-        for message_id in event["attachedMessageIds"]:
+        for message_id in EVENT_FIELDS_RESOLVER.get_attached_messages_ids(event):
             result[message_id].append(event)
 
     return result
@@ -245,7 +250,6 @@ def get_attached_message_ids_index(events: List[Dict]) -> Dict[str, list]:
 # Because it's the same we can you get_category_totals inside
 #
 # STREAMING
-# TODO - NOT-READY -- event["successful"] should be updated by resolver
 def get_type_totals(events: List[Dict]) -> Dict[str, int]:
     """Returns dictionary quantities of events for different event types.
 
@@ -264,8 +268,8 @@ def get_type_totals(events: List[Dict]) -> Dict[str, int]:
     """
     event_types = defaultdict(int)
     for event in events:
-        status = " [ok]" if event["successful"] else " [fail] "
-        event_type = event["eventType"] + status
+        status = " [ok]" if EVENT_FIELDS_RESOLVER.get_status(event) else " [fail] "
+        event_type = EVENT_FIELDS_RESOLVER.get_type(event) + status
         event_types[event_type] += 1
 
     return event_types
@@ -306,8 +310,8 @@ def get_some(
     limit = start + count
     counter = 0
     for event in events:
-        if event["eventType"] == event_type:
-            if failed and event["successful"]:
+        if EVENT_FIELDS_RESOLVER.get_type(event) == event_type:
+            if failed and EVENT_FIELDS_RESOLVER.get_status(event):
                 continue
             if counter >= start:
                 result.append(event)
@@ -340,10 +344,10 @@ def get_related_events(events: List[Dict], messages: List[Dict], count: int) -> 
             ]
     """
     result = []
-    msg_ids = set(message["messageId"] for message in messages)
+    msg_ids = set(MESSAGE_FIELDS_RESOLVER.get_id(message) for message in messages)
 
     for event in events:
-        for msg_id in event["attachedMessageIds"]:
+        for msg_id in EVENT_FIELDS_RESOLVER.get_attached_messages_ids(event):
             if msg_id in msg_ids:
                 result.append(event)
                 if len(result) == count:
@@ -381,7 +385,7 @@ def get_events_by_category(
     counter = 0
     for event in events:
         if categorizer(event) == category:
-            if failed and event["successful"]:
+            if failed and EVENT_FIELDS_RESOLVER.get_status(event):
                 continue
             if counter >= start:
                 result.append(event)
@@ -409,7 +413,7 @@ def get_roots(events: List[Dict], count: int, start: int = 0) -> List[Dict]:
     limit = start + count
     counter = 0
     for event in events:
-        if event["parentEventId"] is None:
+        if EVENT_FIELDS_RESOLVER.get_parent_id(event) is None:
             if counter >= start:
                 result.append(event)
             counter += 1
@@ -438,8 +442,8 @@ def get_parents(events: List[Dict], children: List[Dict]) -> List[Dict]:
                 ...
             ]
     """
-    parent_ids = set(child["parentEventId"] for child in children)
-    return [event for event in events if event["eventId"] in parent_ids]
+    parent_ids = set(EVENT_FIELDS_RESOLVER.get_parent_id(child) for child in children)
+    return [event for event in events if EVENT_FIELDS_RESOLVER.get_id(event) in parent_ids]
 
 
 # NOT STREAMING
@@ -467,9 +471,9 @@ def get_children_from_parent_id(events: List[Dict], parent_id: str, max_events: 
     resolved_parent = {}
     counter = 0
     for event in events:
-        if event["eventId"] == parent_id:
+        if EVENT_FIELDS_RESOLVER.get_id(event) == parent_id:
             resolved_parent = event
-        if event["parentEventId"] == parent_id:
+        if EVENT_FIELDS_RESOLVER.get_parent_id(event) == parent_id:
             children.append(event)
             counter += 1
             if counter == max_events:
@@ -503,10 +507,10 @@ def get_children_from_parents(events: List[Dict], parents: List[Dict], max_event
                 child_events_count
             )
     """
-    result = {parent["eventId"]: [] for parent in parents}
+    result = {EVENT_FIELDS_RESOLVER.get_id(parent): [] for parent in parents}
     events_count = 0
     for event in events:
-        parent_id = event["parentEventId"]
+        parent_id = EVENT_FIELDS_RESOLVER.get_parent_id(event)
         if parent_id not in result:
             continue
         if len(result[parent_id]) < max_events:
@@ -539,11 +543,11 @@ def get_children_from_parents_as_list(events: List[Dict], parents: List[Dict], m
             ]
 
     """
-    parent_ids = set(parent["eventId"] for parent in parents)
+    parent_ids = set(EVENT_FIELDS_RESOLVER.get_id(parent) for parent in parents)
     result = []
     parents_counts = defaultdict(int)
     for event in events:
-        parent_id = event["parentEventId"]
+        parent_id = EVENT_FIELDS_RESOLVER.get_parent_id(event)
         if parent_id not in parent_ids:
             continue
         if parents_counts[parent_id] < max_events:
@@ -578,7 +582,7 @@ def sublist(events: List[Dict], start_time: datetime, end_time: datetime) -> Lis
     start_time = datetime.timestamp(start_time)
     end_time = datetime.timestamp(end_time)
     for event in events:
-        event_time = event["startTimestamp"]["epochSecond"]
+        event_time = EVENT_FIELDS_RESOLVER.get_start_timestamp(event)["epochSecond"]
         if start_time <= event_time <= end_time:
             result.append(event)
 
@@ -618,8 +622,10 @@ def build_roots_cache(events: List[Dict], depth: int, max_level: int) -> Dict:
     prev_levels = get_roots(events, max_level)
     print(f"Level {level}: {len(prev_levels)} events")
     for prev_level in prev_levels:
-        result[prev_level["eventId"]] = {"eventName": prev_level["eventName"], "eventPath": prev_level["eventName"]}
-
+        result[EVENT_FIELDS_RESOLVER.get_id(prev_level)] = {
+            "eventName": EVENT_FIELDS_RESOLVER.get_name(prev_level),
+            "eventPath": EVENT_FIELDS_RESOLVER.get_name(prev_level),
+        }
     while level < depth:
         next_levels = get_children_from_parents_as_list(events, prev_levels, max_level)
         next_levels_count = len(next_levels)
@@ -628,11 +634,11 @@ def build_roots_cache(events: List[Dict], depth: int, max_level: int) -> Dict:
         level += 1
         print(f"Level {level}: {next_levels_count} events")
         for next_level in next_levels:
-            event_id = next_level["eventId"]
-            parent_id = next_level["parentEventId"]
+            event_id = EVENT_FIELDS_RESOLVER.get_id(next_level)
+            parent_id = EVENT_FIELDS_RESOLVER.get_parent_id(next_level)
             result[event_id] = {
-                "eventName": next_level["eventName"],
-                "eventPath": result[parent_id]["eventPath"] + "/" + next_level["eventName"],
+                "eventName": EVENT_FIELDS_RESOLVER.get_name(next_level),
+                "eventPath": result[parent_id]["eventPath"] + "/" + EVENT_FIELDS_RESOLVER.get_name(next_level),
             }
         prev_levels = next_levels
 
@@ -649,7 +655,7 @@ def extract_start_timestamp(event: Dict) -> str:
     Returns:
         str
     """
-    return misc_utils.extract_timestamp(event["startTimestamp"])
+    return misc_utils.extract_timestamp(EVENT_FIELDS_RESOLVER.get_start_timestamp(event))
 
 
 # STREAMING
@@ -693,12 +699,12 @@ def print_event(event: Dict) -> None:
 
     """
     print(
-        f"{extract_start_timestamp(event)} > [{'ok' if event['successful'] else 'fail'}] "
-        f"Type: {event['eventType']} "
-        f"Name: {event['eventName']} "
-        f"ID: {event['eventId']} "
-        f"Parent:{event['parentEventId']} "
-        f"Body:{event['body']}"
+        f"{extract_start_timestamp(event)} > [{'ok' if EVENT_FIELDS_RESOLVER.get_status(event) else 'fail'}] "
+        f"Type: {EVENT_FIELDS_RESOLVER.get_type(event)} "
+        f"Name: {EVENT_FIELDS_RESOLVER.get_name(event)} "
+        f"ID: {EVENT_FIELDS_RESOLVER.get_id(event)} "
+        f"Parent:{EVENT_FIELDS_RESOLVER.get_parent_id(event)} "
+        f"Body:{EVENT_FIELDS_RESOLVER.get_body(event)}"
     )
 
 
@@ -866,8 +872,8 @@ def print_children_from_parents(events: List[Dict], parents: List[Dict], max_eve
     print(f"####### Retrieved Children: {count}")
     for parent in parents:
         print(parent)
-        print(f">>>>>>> Children: {len(tree[parent['eventId']])}")
-        for child in tree[parent["eventId"]]:
+        print(f">>>>>>> Children: {len(tree[EVENT_FIELDS_RESOLVER.get_id(parent)])}")
+        for child in tree[EVENT_FIELDS_RESOLVER.get_id(parent)]:
             print_event(child)
         print(">>>>>>>>>>>>>>>>>>>>>>>>>>>>>>")
 
@@ -890,22 +896,22 @@ def print_children_stats_from_parents(
     for parent in parents:
         start_timestamp, end_timestamp = 0, 0
         events_passed, events_failed = 0, 0
-        for child in tree[parent["eventId"]]:
-            child_start_time_epoch = child["startTimestamp"]["epochSecond"]
+        for child in tree[EVENT_FIELDS_RESOLVER.get_id(parent)]:
+            child_start_time_epoch = EVENT_FIELDS_RESOLVER.get_start_timestamp(child)["epochSecond"]
             if child_start_time_epoch > end_timestamp:
                 end_timestamp = child_start_time_epoch
             if start_timestamp == 0 or child_start_time_epoch < start_timestamp:
                 start_timestamp = child_start_time_epoch
 
-            if child["successful"]:
+            if EVENT_FIELDS_RESOLVER.get_status(child):
                 events_passed += 1
             else:
                 events_failed += 1
 
         table.append(
             [
-                parent["eventName"],
-                "[ok]" if parent["successful"] else "[fail]",
+                EVENT_FIELDS_RESOLVER.get_name(parent),
+                "[ok]" if EVENT_FIELDS_RESOLVER.get_status(parent) else "[fail]",
                 events_passed,
                 events_failed,
                 datetime.fromtimestamp(start_timestamp).isoformat(),